menambahkan fungsi confirm pada fitur hapus data

+++ main/daftar_pemesan.php
@@ -87,7 +87,7 @@ include "../lib/koneksi.php"
                                 <td>Rp " . number_format($row['total_tagihan'], 0, ',', '.') . "</td>
                                 <td>
                                     <a href='detail_pemesan.php?id={$row['id']}' class='btn btn-info btn-sm'>Lihat Detail</a>
-                                    <a href='../lib/hapus.php?id={$row['id']}' class='btn btn-danger btn-sm'>Hapus</a>
+                                    <a href='../lib/hapus.php?id={$row['id']}' class='btn btn-danger btn-sm' onclick='return confirm(\"Apakah Anda yakin ingin menghapus data ini?\")'>Hapus</a>
                                 </td>
                             </tr>";
                             $no++;
